requirements.txt
llama-cpp-python>=0.2.20
huggingface-hub>=0.19.0
transformers>=4.48.0
torch
accelerate
safetensors
playwright>=1.40.0
langchain>=0.1.0
langchain-community>=0.0.10
reportlab>=4.0.0
Pillow>=10.0.0
